# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RC Patel Interior API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type CreateEnquiryBodyProjectType =
  (typeof CreateEnquiryBodyProjectType)[keyof typeof CreateEnquiryBodyProjectType];

export const CreateEnquiryBodyProjectType = {
  residential: "residential",
  commercial: "commercial",
  office: "office",
  hospitality: "hospitality",
  other: "other",
} as const;

export interface CreateEnquiryBody {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  projectType: CreateEnquiryBodyProjectType;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  budget?: string | null;
  message: string;
  imageObjectPaths?: string[];
}

export type UpdateEnquiryStatusBodyStatus =
  (typeof UpdateEnquiryStatusBodyStatus)[keyof typeof UpdateEnquiryStatusBodyStatus];

export const UpdateEnquiryStatusBodyStatus = {
  new: "new",
  contacted: "contacted",
  in_progress: "in_progress",
  completed: "completed",
  closed: "closed",
} as const;

export interface UpdateEnquiryStatusBody {
  status: UpdateEnquiryStatusBodyStatus;
}

export interface Enquiry {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  projectType: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  budget?: string | null;
  message: string;
  status: string;
  imageObjectPaths: string[];
  createdAt: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListEnquiriesParams = {
  adminKey: string;
};

export type GetEnquiryParams = {
  adminKey: string;
};

export type UpdateEnquiryStatusParams = {
  adminKey: string;
};