# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RC Patel Interior API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: enquiries
    description: Enquiry management
  - name: storage
    description: Object storage operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /enquiries:
    post:
      operationId: createEnquiry
      tags: [enquiries]
      summary: Submit a new enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnquiryBody"
      responses:
        "201":
          description: Enquiry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listEnquiries
      tags: [enquiries]
      summary: List all enquiries (admin)
      parameters:
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of enquiries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enquiry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /enquiries/{id}:
    get:
      operationId: getEnquiry
      tags: [enquiries]
      summary: Get a single enquiry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Enquiry detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEnquiryStatus
      tags: [enquiries]
      summary: Update enquiry status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: adminKey
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnquiryStatusBody"
      responses:
        "200":
          description: Updated enquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enquiry"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"

  /storage/objects/{objectPath}:
    get:
      operationId: getObject
      tags: [storage]
      summary: Serve a stored object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    CreateEnquiryBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        projectType:
          type: string
          enum: [residential, commercial, office, hospitality, other]
        city:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        message:
          type: string
        imageObjectPaths:
          type: array
          items:
            type: string
      required:
        - name
        - phone
        - projectType
        - message

    UpdateEnquiryStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [new, contacted, in_progress, completed, closed]
      required:
        - status

    Enquiry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        projectType:
          type: string
        city:
          type: ["string", "null"]
        budget:
          type: ["string", "null"]
        message:
          type: string
        status:
          type: string
        imageObjectPaths:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - projectType
        - message
        - status
        - imageObjectPaths
        - createdAt

    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType

    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath