# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RC Patel Interior API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Submit a new enquiry
 */
export const CreateEnquiryBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
  email: zod.string().nullish(),
  projectType: zod.enum([
    "residential",
    "commercial",
    "office",
    "hospitality",
    "other",
  ]),
  city: zod.string().nullish(),
  budget: zod.string().nullish(),
  message: zod.string(),
  imageObjectPaths: zod.array(zod.string()).optional(),
});

/**
 * @summary List all enquiries (admin)
 */
export const ListEnquiriesQueryParams = zod.object({
  adminKey: zod.coerce.string(),
});

export const ListEnquiriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  email: zod.string().nullish(),
  projectType: zod.string(),
  city: zod.string().nullish(),
  budget: zod.string().nullish(),
  message: zod.string(),
  status: zod.string(),
  imageObjectPaths: zod.array(zod.string()),
  createdAt: zod.string(),
});
export const ListEnquiriesResponse = zod.array(ListEnquiriesResponseItem);

/**
 * @summary Get a single enquiry (admin)
 */
export const GetEnquiryParams = zod.object({
  id: zod.coerce.number(),
});

export const GetEnquiryQueryParams = zod.object({
  adminKey: zod.coerce.string(),
});

export const GetEnquiryResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  email: zod.string().nullish(),
  projectType: zod.string(),
  city: zod.string().nullish(),
  budget: zod.string().nullish(),
  message: zod.string(),
  status: zod.string(),
  imageObjectPaths: zod.array(zod.string()),
  createdAt: zod.string(),
});

/**
 * @summary Update enquiry status (admin)
 */
export const UpdateEnquiryStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateEnquiryStatusQueryParams = zod.object({
  adminKey: zod.coerce.string(),
});

export const UpdateEnquiryStatusBody = zod.object({
  status: zod.enum(["new", "contacted", "in_progress", "completed", "closed"]),
});

export const UpdateEnquiryStatusResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  email: zod.string().nullish(),
  projectType: zod.string(),
  city: zod.string().nullish(),
  budget: zod.string().nullish(),
  message: zod.string(),
  status: zod.string(),
  imageObjectPaths: zod.array(zod.string()),
  createdAt: zod.string(),
});

/**
 * @summary Request a presigned upload URL
 */
export const RequestUploadUrlBody = zod.object({
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string(),
  objectPath: zod.string(),
});

/**
 * @summary Serve a stored object
 */
export const GetObjectParams = zod.object({
  objectPath: zod.coerce.string(),
});