# lib/api-zod/src/generated/types/createEnquiryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RC Patel Interior API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateEnquiryBodyProjectType } from "./createEnquiryBodyProjectType";

export interface CreateEnquiryBody {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  projectType: CreateEnquiryBodyProjectType;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  budget?: string | null;
  message: string;
  imageObjectPaths?: string[];
}