# lib/api-zod/src/generated/types/enquiry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RC Patel Interior API
 * OpenAPI spec version: 0.1.0
 */

export interface Enquiry {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  projectType: string;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  budget?: string | null;
  message: string;
  status: string;
  imageObjectPaths: string[];
  createdAt: string;
}