# lib/api-zod/src/generated/types/updateEnquiryStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RC Patel Interior API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEnquiryStatusBodyStatus } from "./updateEnquiryStatusBodyStatus";

export interface UpdateEnquiryStatusBody {
  status: UpdateEnquiryStatusBodyStatus;
}